requirements.txt
Flask
Werkzeug
colorgram.py
gunicorn
